requirements.txt
requests
python-dateutil
celery
redis
django
djangorestframework
django-celery-beat
APScheduler